feat(reset_password): Implement backend for password reset.
User can now request to reset their password. On their request,
user will receive an email with a link to reset their password.
A user's ID and a password reset token are used to validate their
request because they are long string unique to each user. A password
reset token should only be used by a single user and should be hashed
before being stored in the database for security reasons.